+++ src/kwik/typings/sorting_query.py
@@ -1,6 +1,0 @@
-"""Type definitions for query sorting."""
-
-from __future__ import annotations
-
-SortingQuery = str
-ParsedSortingQuery = list[tuple[str, str]]